Separated the classes and the macros from each other Added comments
