+++ main.py
@@ -1,99 +1,24 @@
-import pygame
+'''
+Contains all the information related to the TicTacToe class
+The game starts here
+'''
 
-# (0,0) = Top Left Corner
-SCREEN_WIDTH = 1000
-SCREEN_HEIGHT = 600
-window_size = (SCREEN_WIDTH, SCREEN_HEIGHT)
-screen = pygame.display.set_mode(window_size)
+from Table import *
+from Macros import *
 
-GAME_NAME = "TIC TAC TOE"
-pygame.display.set_caption(GAME_NAME)
-
-background = pygame.Surface(screen.get_size())
-background_image = pygame.image.load("images/Bamboo_bc_1.bmp").convert()
-x_mark = pygame.image.load("images/Player_X.png").convert_alpha()
-o_mark = pygame.image.load("images/Player_O.png").convert_alpha()
-
-class Tiles(pygame.sprite.DirtySprite):
-    def __init__(self, new_image, x, y):
-        # Call the parent class (Sprite) constructor
-        pygame.sprite.Sprite.__init__(self)
-        self.image = new_image
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.x = y
-
-    def update(self):
-        self.rect.x += 50
-        self.rect.x += 50
-
-# Tiles as pieces on the board - if I click on a certain Tile object it will change accordingly
-class Board:
-    def __init__(self):
-        tile_width = 200
-        tile_height = 200
-
-        # First Row
-        self.tile1 = pygame.Rect(200, 0, tile_width, tile_height)
-        self.tile2 = pygame.Rect(400, 0, tile_width, tile_height)
-        self.tile3 = pygame.Rect(600, 0, tile_width, tile_height)
-        # Second Row
-        self.tile4 = pygame.Rect(200, 200, tile_width, tile_height)
-        self.tile5 = pygame.Rect(400, 200, tile_width, tile_height)
-        self.tile6 = pygame.Rect(600, 200, tile_width, tile_height)
-        # Third Row
-        self.tile7 = pygame.Rect(200, 400, tile_width, tile_height)
-        self.tile8 = pygame.Rect(400, 400, tile_width, tile_height)
-        self.tile9 = pygame.Rect(600, 400, tile_width, tile_height)
-
-        self.tiles_on_board = [
-            self.tile1, self.tile2, self.tile3,
-            self.tile4, self.tile5, self.tile6,
-            self.tile7, self.tile8, self.tile9
-        ]
-
-        self.tiles = {
-            # First Row
-            (200, 0): (0, 0),
-            (400, 0): (0, 1),
-            (600, 0): (0, 2),
-
-            # Second Row
-            (200, 200): (1, 0),
-            (400, 200): (1, 1),
-            (600, 200): (1, 2),
-
-            # Third Row
-            (200, 400): (2, 0),
-            (400, 400): (2, 1),
-            (600, 400): (2, 2),
-        }
 
 class TicTacToe:
     def __init__(self):
+        """
+        Initializes the table, players
+        """
+
         self.board = Board()
         self.tic_tac_table = [
             ["", "", ""],
             ["", "", ""],
             ["", "", ""]
         ]
-
-        self.tiles = {
-            # First Row
-            (0, 0): (200, 0),
-            (0, 1): (400, 0),
-            (0, 2): (600, 0),
-
-            # Second Row
-            (1, 0): (200, 200),
-            (1, 1): (400, 200),
-            (1, 2): (600, 200),
-
-            # Third Row
-            (2, 0): (200, 400),
-            (2, 1): (400, 400),
-            (2, 2): (600, 400),
-        }
 
         self.player1 = {
             "name": "Player X",
@@ -121,67 +46,55 @@ class TicTacToe:
                f"{self.tic_tac_table[2][:]}\n" \
                f"\n\n\n\n"
 
+    def update_screen(self):
+        """
+        Redraws the screen after every turn
+        """
 
-    def draw_lines(self):
-        # -- update the table
-        line_color = pygame.Color(0, 0, 0)
-        # Vertical
-        pygame.draw.line(background, line_color, (200, 0), (200, 600), 10)
-        pygame.draw.line(background, line_color, (400, 0), (400, 600), 10)
-        pygame.draw.line(background, line_color, (600, 0), (600, 600), 10)
-        pygame.draw.line(background, line_color, (800, 0), (800, 600), 10)
-
-        # Horizontal
-        pygame.draw.line(background, line_color, (200, 0), (800, 0), 10)
-        pygame.draw.line(background, line_color, (200, 200), (800, 200), 10)
-        pygame.draw.line(background, line_color, (200, 400), (800, 400), 10)
-        pygame.draw.line(background, line_color, (200, 600), (800, 600), 10)
-
-    def draw_table(self):
         background.blit(background_image, (0, 0))
-
-        self.draw_lines()
-
-        screen.blit(background, (0, 0))
-        pygame.display.update()
-
-    def update_screen(self, tile):
-        background.blit(background_image, (0, 0))
-        self.draw_lines()
+        self.board.draw_lines()
 
         for coord in self.player1["occupied_tiles"]:
             if coord:
-                background.blit(self.player1["image"], self.tiles[coord])
+                background.blit(self.player1["image"], self.board.tiles[coord])
         for coord in self.player2["occupied_tiles"]:
             if coord:
-                background.blit(self.player2["image"], self.tiles[coord])
+                background.blit(self.player2["image"], self.board.tiles[coord])
 
         screen.blit(background, (0, 0))
         pygame.display.update()
 
     def update_player(self, tile):
-        #If the cursor is on the Board
+        """
+        Updates the data of a player if the player did a regular move
+        :param tile: one of the table's tile top left coordinates
+        """
+
+        # If the cursor is on the Board
         if tile in self.board.tiles:
             board_coord = self.board.tiles[tile]
 
-            #If there is no mark on the Tile
+            # If there is no mark on the Tile
             if not self.tic_tac_table[board_coord[0]][board_coord[1]]:
                 player = self.players[0] if self.player1_turn else self.players[1]
                 self.tic_tac_table[board_coord[0]][board_coord[1]] = player["mark"]
 
-                #If we start to override the players' marks
+                # If we start to override the players' marks
                 if player["occupied_tiles"][player["turns"]]:
                     fourth_mark = player["occupied_tiles"][player["turns"]]
                     self.tic_tac_table[fourth_mark[0]][fourth_mark[1]] = ""
-                    # CHANGE THIS MARK POSITION TO THE NEW ONE
 
                 player["occupied_tiles"][player["turns"]] = board_coord
                 player["turns"] = (player["turns"] + 1) % 3
 
-                self.update_screen(tile)
+                self.update_screen()
                 self.player1_turn = not self.player1_turn
 
     def check_winner(self):
+        """
+        The game checks if one of the player wins by checking if the rows or columns have the same character
+        """
+
         # First Row
         if self.tic_tac_table[0][0] == self.tic_tac_table[0][1] == self.tic_tac_table[0][2] != "":
             print("The winner is:", self.tic_tac_table[0][0])
@@ -210,6 +123,11 @@ class TicTacToe:
             print("The winner is:", self.tic_tac_table[0][2])
 
     def check_mouse_coords(self):
+        """
+        Checks if the player clicked on a tile on the table
+        :return: the tile's top left coordinate
+        """
+
         pygame.time.delay(100)
 
         # Mouse Action
@@ -218,9 +136,11 @@ class TicTacToe:
             if tile.collidepoint(mouse_coord):
                 return tile.topleft
 
-
     def start_game(self):
-        game.draw_table()
+        """
+        The game loop starts here and stops when the player quits the game
+        """
+        self.board.draw_table()
 
         while self.running:
             for event in pygame.event.get():
@@ -243,7 +163,7 @@ if __name__ == '__main__':
     game = TicTacToe()
     game.start_game()
 
-'''
+"""
 # MENU
 # Keyboard Action
 elif event.type == pygame.KEYDOWN:
@@ -255,4 +175,4 @@ elif event.type == pygame.KEYDOWN:
     elif keys[pygame.K_UP]:
 
     elif keys[pygame.K_DOWN]:
-'''
+"""
